# Fogadoora Backend/src/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "REST API minta backend",
        "version": "1.0.0",
        "description": "<img alt='HaHu' height='400px' src='https://nitslaszlo.github.io/static/hahu/Hahu.jpg' />"
    },
    "servers": [
        {
            "url": "http://localhost:5000"
        },
        {
            "url": "https://jedlik.cyclic.app"
        }
    ],
    "paths": {
...
        "/api/teachers": {
            "get": {
                "tags": ["teachers"],
                "summary": "Összes tanár lekérdezése.",
                "description": "Az összes tanár adatát tudjuk lekérdezni.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/teacher"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error.message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["teachers"],
                "summary": "Új tanár mentése.",
                "description": "Az új tanár mentése a teachers táblában",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "$ref": "#/components/schemas/teacher"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK (Az új, elmentett tanár tér vissza.)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/teacher"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error.message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/teachers/{offset}/{limit}/{field}": {
            "get": {
                "tags": ["teachers"],
                "summary": "Paginált adatok lekérése.",
                "description": "Tanárok lekérdezése lapozáshoz.",
                "parameters": [
                    {
                        "name": "offset",
                        "description": "Ennyi rekordot hagyunk el (skip) a lekérdezés 'elejéről'.",
                        "in": "path",
                        "schema": {
                            "type": "number",
                            "example": "5"
                        },
                        "required": true
                    },
                    {
                        "name": "limit",
                        "description": "Ennyi rekordot küld (take) maximum az API",
                        "in": "path",
                        "schema": {
                            "type": "number",
                            "example": "5"
                        },
                        "required": true
                    },
                    {
                        "name": "field",
                        "description": "Melyik mező szerint legyenek rendezve a rekordok. A kötőjel '-' karakter a csökkenő rendezést jelöli.",
                        "in": "path",
                        "schema": {
                            "type": "string",
                            "example": "-cim"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "count": {
                                            "type": "number",
                                            "description": "Az összes rekord száma.",
                                            "example": "16"
                                        },
                                        "data": {
                                            "type": "array",
                                            "description": "A maximum {limit} számú rekord.",
                                            "items": {
                                                "$ref": "#/components/schemas/teacher"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error.message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/teachers/{offset}/{limit}/{field}/{filter}": {
            "get": {
                "tags": ["teachers"],
                "summary": "Paginált adatok lekérése szűréssel.",
                "description": "Tanárok lekérdezése lapozáshoz szűréssel.",
                "parameters": [
                    {
                        "name": "offset",
                        "description": "Ennyi rekordot hagyunk el (skip) a lekérdezés 'elejéről'.",
                        "in": "path",
                        "schema": {
                            "type": "number",
                            "example": "0"
                        },
                        "required": true
                    },
                    {
                        "name": "limit",
                        "description": "Ennyi rekordot küld (take) maximum az API.",
                        "in": "path",
                        "schema": {
                            "type": "number",
                            "example": "2"
                        },
                        "required": true
                    },
                    {
                        "name": "field",
                        "description": "Melyik mező szerint legyenek rendezve a rekordok. A kötőjel '-' karakter a csökkenő rendezést jelöli.",
                        "in": "path",
                        "schema": {
                            "type": "string",
                            "example": "-cim"
                        },
                        "required": true
                    },
                    {
                        "name": "filter",
                        "description": "Szűrő - a 'cim' és a 'leiras' mezőkben keressük.",
                        "in": "path",
                        "schema": {
                            "type": "string",
                            "example": "sport"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "count": {
                                            "type": "number",
                                            "description": "A szűrőnek megfelelő rekordok száma",
                                            "example": "16"
                                        },
                                        "data": {
                                            "type": "array",
                                            "description": "A maximum {limit} számú, szűrőnek megfelelő rekord.",
                                            "items": {
                                                "$ref": "#/components/schemas/teacher"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error.message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "teacher": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "description": "PK - A tanár azonosítója (ObjectID).",
                        "example": "4"
                    },
                    "név": {
                        "type": "string",
                        "description": "Az azonosítóhoz tartozó tanárnak a neve.",
                        "example": "Teszt Elek"
                    }
                }
            }
        }
    }
}